requirements.txt
Django==4.2.19
djangorestframework==3.15.2
Markdown==3.7
django-filter==24.3
django-rest-knox==4.2.0
django_extensions==3.2.3
drf_standardized_errors==0.14.1
django-cors-headers==4.7.0
django-cryptography==1.1
djangorestframework-recursive==0.1.2
pillow==11.1.0
drf-spectacular==0.28.0
django-environ==0.12.0